requirement.txt
Package           Version
----------------- -------
annotated-types   0.7.0
anyio             4.5.2
click             8.1.8
dnspython         2.6.1
exceptiongroup    1.2.2
fastapi           0.115.9
h11               0.14.0
idna              3.10
motor             3.6.1
pip               23.0.1
pydantic          2.10.6
pydantic_core     2.27.2
pymongo           4.9.2
python-dotenv     1.0.1
setuptools        56.0.0
sniffio           1.3.1
starlette         0.44.0
typing_extensions 4.12.2
uvicorn           0.33.0
